pyproject.toml
[project]
name = "leethack"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "black>=25.1.0",
    "boto3>=1.38.27",
    "dj-database-url>=3.0.0",
    "django>=5.2.1",
    "django-debug-toolbar>=5.2.0",
    "django-environ>=0.12.0",
    "django-filter>=25.1",
    "django-storages>=1.14.6",
    "djangorestframework>=3.16.0",
    "drf-spectacular>=0.28.0",
    "factory-boy>=3.3.3",
    "gunicorn>=23.0.0",
    "numpy>=2.2.6",
    "pillow>=11.2.1",
    "psycopg2>=2.9.10",
    "pytest-django>=4.11.1",
    "whitenoise>=6.9.0",
]
